refactor(app): move middleware to package
